move to separate addon, add more actions
