+++ addons/misc/XEH_PREP.hpp
@@ -1,12 +1,8 @@
 PREP(addIVbag);
 PREP(addVehicleCarryLoadActions);
 PREP(addVehicleUnloadCarryPatientActions);
-PREP(attachHeliStretcher);
-PREP(attachStretcher);
 PREP(BubbleWrap);
-PREP(canAdjustRopes);
 PREP(canCheckDogtag);
-PREP(canDeployStretcher);
 PREP(canStitchFullBody);
 PREP(carryObject);
 PREP(checkArmbandCondition);
@@ -14,7 +10,6 @@ PREP(conditionExecutionDeath);
 PREP(conditionIV);
 PREP(conditionIVstand);
 PREP(conditionSecondChance);
-PREP(deployStretcher);
 PREP(FAK_ammoToArray);
 PREP(FAK_arrayToAmmo);
 PREP(FAK_checkRepack);
@@ -34,14 +29,10 @@ PREP(handleUnconscious);
 PREP(hasStableVitals);
 PREP(incompatibilityWarning);
 PREP(init);
-PREP(lowerStretcher);
-PREP(raiseStretcher);
 PREP(removeIVbag);
-PREP(retractStretcher);
 PREP(setUnconscious);
 PREP(slingArmband);
 PREP(stopCarryingPrompt);
-PREP(stretcher);
 PREP(stringToArray);
 PREP(surgicalKitProgressFullBody);
 PREP(switchMagItems);
