+++ addons/stretcher/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\kat\addons\stretcher
