+++ addons/stretcher/functions/fnc_attachHeliStretcher.sqf
@@ -1,19 +1,25 @@
 #include "..\script_component.hpp"
 
 /*
- * Author: Battlekeeper, modified by YetheSamartaka and Tomcat
- * Handles the behaviour of Helistretcher - attaching it and using it.
+ * Author: MiszczuZPolski
+ * Attaches helistretcher to helicopter
  *
+ * Arguments:
+ * 0: Player <OBJECT>
+ * 1: Helicopter <OBJECT>
  *
  * Example:
- * [_unit,_item] call kat_misc_fnc_heliStretcherAttach;
+ * [player, cursorObject] call kat_stretcher_fnc_attachHeliStretcher;
  *
  * Public: No
  */
 
 params ["_player","_target"];
 
+_player removeItem "kat_HeliStretcher";
 private _stretcher = createVehicle ["kat_stretcher", (_target modelToWorld [0,0,-5])];
 _stretcher attachTo [_target, [0,0,-.1], "slingload0"];
 _target setVariable [QGVAR(isStretcherAttached), true, true];
 _target setVariable [QGVAR(stretcher), _stretcher, true];
+
+
