+++ addons/stretcher/functions/fnc_canDeployStretcher.sqf
@@ -1,15 +1,18 @@
 #include "..\script_component.hpp"
 
 /*
- * Author: Battlekeeper, modified by YetheSamartaka and Tomcat
- * Handles the behaviour of Helistretcher - attaching it and using it.
+ * Author: MiszczuZPolski
+ * Checks if attached helistretcher can be deployed
  *
+ * Arguments:
+ * 0: Player <OBJECT>
  *
  * Example:
- * [_unit,_item] call kat_misc_fnc_heliStretcherAttach;
+ * [player] call kat_stretcher_fnc_canDeployStretcher;
  *
  * Public: No
  */
 
 params ["_player"];
+
 objectParent _player getVariable [QGVAR(isStretcherAttached), false] && !(objectParent _player getVariable [QGVAR(stretcherDeployed), false])
