+++ addons/stretcher/functions/fnc_deployStretcher.sqf
@@ -1,12 +1,15 @@
 #include "..\script_component.hpp"
 
 /*
- * Author: Battlekeeper, modified by YetheSamartaka and Tomcat
- * Handles the behaviour of Helistretcher - attaching it and using it.
+ * Author: MiszczuZPolski
+ * Deploys helistretcher attached below helicopter
  *
+ * Arguments:
+ * 0: Player <OBJECT>
+ * 1: Helicopter <OBJECT>
  *
  * Example:
- * [_unit,_item] call kat_misc_fnc_heliStretcherAttach;
+ * [player, cursorObject] call kat_stretcher_fnc_deployStretcher;
  *
  * Public: No
  */
@@ -21,10 +24,10 @@ private _rope2 = ropeCreate [_target, "slingload0", _stretcher, [-.25,1,0], 5];
 private _rope3 = ropeCreate [_target, "slingload0", _stretcher, [-.25,-1,0], 5];
 private _rope4 = ropeCreate [_target, "slingload0", _stretcher, [.25,-1,0], 5];
 
-ropeUnwind [_rope1, 2, 5];
-ropeUnwind [_rope2, 2, 5];
-ropeUnwind [_rope3, 2, 5];
-ropeUnwind [_rope4, 2, 5];
+ropeUnwind [_rope1, 2, 10];
+ropeUnwind [_rope2, 2, 10];
+ropeUnwind [_rope3, 2, 10];
+ropeUnwind [_rope4, 2, 10];
 
 _target setVariable [QGVAR(stretcherDeployed), true, true];
 _target setVariable [QGVAR(ropes), [_rope1,_rope2,_rope3,_rope4], true];
