+++ addons/stretcher/functions/fnc_lowerStretcher.sqf
@@ -1,12 +1,15 @@
 #include "..\script_component.hpp"
 
 /*
- * Author: Battlekeeper, modified by YetheSamartaka and Tomcat
- * Handles the behaviour of Helistretcher - attaching it and using it.
+ * Author: MiszczuZPolski
+ * Lowers helistretcher by 10m
  *
+ * Arguments:
+ * 0: Player <OBJECT>
+ * 1: Helicopter <OBJECT>
  *
  * Example:
- * [_unit,_item] call kat_misc_fnc_heliStretcherAttach;
+ * [player, cursorObject] call kat_stretcher_fnc_lowerStretcher;
  *
  * Public: No
  */
