+++ addons/stretcher/models/stretcher/seat_destruct.rvmat
@@ -8,7 +8,7 @@ PixelShaderID = "NormalMapMacroASSpecularDIMap";
 VertexShaderID = "NormalMapAS";
 
 class Stage1 {
-    texture = "x\kat\addons\misc\models\stretcher\seat_nohq.paa";
+    texture = "x\kat\addons\stretcher\models\stretcher\seat_nohq.paa";
     uvSource = "tex";
 
     class uvTransform {
@@ -20,7 +20,7 @@ class Stage1 {
 };
 
 class Stage2 {
-    texture = "x\kat\addons\misc\models\stretcher\seat_destruct_mc.paa";
+    texture = "x\kat\addons\stretcher\models\stretcher\seat_destruct_mc.paa";
     uvSource = "tex";
 
     class uvTransform {
@@ -44,7 +44,7 @@ class Stage3 {
 };
 
 class Stage4 {
-    texture = "x\kat\addons\misc\models\stretcher\seat_destruct_smdi.paa";
+    texture = "x\kat\addons\stretcher\models\stretcher\seat_destruct_smdi.paa";
     uvSource = "tex";
 
     class uvTransform {
