+++ addons/stretcher/models/stretcher/seat_destruct_mc.paa
