+++ addons/stretcher/models/stretcher/seat_destruct_smdi.paa
