+++ addons/stretcher/models/stretcher/seat_nohq.paa
