+++ addons/stretcher/models/stretcher/seat_smdi.paa
