+++ addons/stretcher/ui/uparrow.paa
